contract-generator-gateway/requirements.txt
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.8.3
django==4.2.23
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.10
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
typing-extensions==4.13.2
uritemplate==4.1.1
